requirements.txt
blinker==1.8.2
certifi==2024.6.2        
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Login==0.6.3       
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysql-connector-python==8.4.0
packaging==24.1
pluggy==1.5.0
pytest==8.2.2
requests==2.32.3
urllib3==2.2.2
Werkzeug==3.0.3
